# docs/openapi/hrportal-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HR Portal API",
    "description": "Modular monolith HR platform API. Business endpoints require Bearer JWT and X-Tenant-Id header.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/leave-requests": {
      "get": {
        "tags": [
          "Leave"
        ],
        "summary": "List all leave requests.",
        "description": "Auth: ManagerOrAbove",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaveRequestDto"
                  }
                },
                "example": [
                  {
                    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "startDate": "2025-07-01",
                    "endDate": "2025-07-05",
                    "type": "Annual",
                    "status": "Pending",
                    "reason": "Summer holiday",
                    "approvedBy": null,
                    "approvedAt": null
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Leave"
        ],
        "summary": "Submit a new leave request.",
        "description": "Auth: Authenticated",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaveRequest"
              },
              "example": {
                "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "startDate": "2025-07-01",
                "endDate": "2025-07-05",
                "type": "Annual",
                "reason": "Summer holiday"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestDto"
                },
                "example": {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "startDate": "2025-07-01",
                  "endDate": "2025-07-05",
                  "type": "Annual",
                  "status": "Pending",
                  "reason": "Summer holiday",
                  "approvedBy": null,
                  "approvedAt": null
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/400",
                  "title": "Bad request",
                  "status": 400,
                  "detail": "Validation failed.",
                  "errorCode": "VALIDATION_ERROR"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{id}": {
      "get": {
        "tags": [
          "Leave"
        ],
        "summary": "Get leave request by ID.",
        "description": "Auth: Authenticated",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestDto"
                },
                "example": {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "startDate": "2025-07-01",
                  "endDate": "2025-07-05",
                  "type": "Annual",
                  "status": "Pending",
                  "reason": "Summer holiday",
                  "approvedBy": null,
                  "approvedAt": null
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/404",
                  "title": "Not found",
                  "status": 404,
                  "detail": "Employee with key '...' was not found.",
                  "errorCode": "NOT_FOUND"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Leave"
        ],
        "summary": "Cancel a pending leave request.",
        "description": "Auth: Authenticated",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/404",
                  "title": "Not found",
                  "status": 404,
                  "detail": "Employee with key '...' was not found.",
                  "errorCode": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{id}/approve": {
      "put": {
        "tags": [
          "Leave"
        ],
        "summary": "Approve a pending leave request.",
        "description": "Auth: ManagerOrAbove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestDto"
                },
                "example": {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "startDate": "2025-07-01",
                  "endDate": "2025-07-05",
                  "type": "Annual",
                  "status": "Pending",
                  "reason": "Summer holiday",
                  "approvedBy": null,
                  "approvedAt": null
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/400",
                  "title": "Bad request",
                  "status": 400,
                  "detail": "Validation failed.",
                  "errorCode": "VALIDATION_ERROR"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/404",
                  "title": "Not found",
                  "status": 404,
                  "detail": "Employee with key '...' was not found.",
                  "errorCode": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{id}/reject": {
      "put": {
        "tags": [
          "Leave"
        ],
        "summary": "Reject a pending leave request.",
        "description": "Auth: ManagerOrAbove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant slug (e.g. demo). Required on all business endpoints.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "demo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeaveRequest"
              },
              "example": {
                "reason": "Insufficient coverage"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeaveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestDto"
                },
                "example": {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "employeeId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "startDate": "2025-07-01",
                  "endDate": "2025-07-05",
                  "type": "Annual",
                  "status": "Pending",
                  "reason": "Summer holiday",
                  "approvedBy": null,
                  "approvedAt": null
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/400",
                  "title": "Bad request",
                  "status": 400,
                  "detail": "Validation failed.",
                  "errorCode": "VALIDATION_ERROR"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/404",
                  "title": "Not found",
                  "status": 404,
                  "detail": "Employee with key '...' was not found.",
                  "errorCode": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — insufficient role for this policy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLeaveRequest": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LeaveRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "approvedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RejectLeaveRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}